Remove redundant casts.

+++ src/main/java/bsh/Primitive.java
@@ -609,7 +609,7 @@ public final class Primitive implements Serializable {
                 return BigDecimal.ONE.setScale(1);
         } else if (toType == BigDecimal.class) {
             if ( number instanceof BigDecimal )
-                return (BigDecimal) number;
+                return number;
             if ( number instanceof BigInteger )
                 return new BigDecimal((BigInteger) number).setScale(1);
             if ( Types.isFloatingpoint(number) )
